Refactor settings editor components
* Enhance the settings editor structure by streamlining component usage and improving code organization.
