+++ packages/js/settings-editor/src/components/info-view/index.tsx
